email_validation/tests: Return anonymous email when inviting existing users.
Previously during an error with invite_users_backend codepath, we returned
3 error items - invited email, error message, is_deactivated. However, for
organizations which haven't enabled EMAIL_ADDRESS_VISIBILITY_EVERYONE the
actual user email stored in the frontend is anonymised and different from
the invited email.

To fix this, we return back the anonymous emails for existing accounts along
with the other parameters. This results in 4 return values ->
    - email (invited email)
    - error_msg
    - is_deactivated
    - maybe_anonymous_email

Incase EMAIL_ADDRESS_VISIBILITY_EVERYONE is enabled for the realm,
maybe_anonymous_email would be same as the invited email. This is also
the case for cross_realm_bots. For invalid emails, we just return back
no strings.

Tests amended.