js: invite: Add UI to subscribe users with existing accounts.
When inviting existing accounts in the realm, the user is presented
with an option to subscribe existing users to the marked streams as
well. The error option displays the users with accounts and the streams
to which they would be subscribed.

The actual stream subscription can result in three cases:
    1. All the users were newly subscribed to the selected streams.
    2. All the users were already part of selected streams.
    3. Some users got newly subscribed, while some others were
       already subscribed.
Appropriate error messages are show.

Fixes #15784.