rendered_markdown: Display view-code-in-playground button in codeblocks.
Clicking on the view-code-in-playground button extracts the code content and
sets the target href combining the url_prefix and the extracted code.

This is more of a prototype and hence the url_prefix is hardcoded in a
dictionary for now. Future steps would be to add support for realm admins
to configure their own choices. Also, the structure of the hardcoded message
is similar to what we want in the eventual configuration API.

Fixes: #11618