realm/playground: Add API endpoint for creating playground entry.
This endpoint will allow clients to create a playground entry containing
the name, pygments language and url_prefix for the playground of their
choice.

Added the `do_*` function in-charge of creating the entry in the model.
However, we aren't handling sending events in this commit, which will be
done in a follow up commit.

Further, added the endpoints to the `pending_endpoints` list to prevent
our openapi tool from flagging these as undocumented.

Tests added.